[Adjustment][Ezra] header coming soon 404
