+++ components/Layout.tsx
@@ -10,7 +10,7 @@ function Layout(props: PropsInterface) {
   const { children } = props;
 
   return (
-    <div className="flex flex-col h-screen">
+    <div className="flex flex-col min-h-screen">
       <Header />
       <div className="flex-1">{children}</div>
       <Footer />
