+++ pages/events.tsx
@@ -1,0 +1,13 @@
+import ComingSoon from "@/components/ComingSoon";
+import Layout from "@/components/Layout";
+import React from "react";
+
+function events() {
+  return (
+    <Layout>
+      <ComingSoon />
+    </Layout>
+  );
+}
+
+export default events;
